+++ src/uci/fen.hpp
@@ -65,7 +65,7 @@ class FEN
          * @param rights Binary notation of the rights (1 for right of castle, else 0).
          * @return A string representing the castling rights of both sides in the game.
          */
-        static std::string castling_rights_to_fen(int rights);
+        static std::string castling_rights_to_fen(uint8_t rights);
 
         /**
          * @brief Gets the FEN notation of the castling rights.
@@ -73,5 +73,5 @@ class FEN
          * @param rights Binary notation of the rights (1 for right of castle, else 0).
          * @return A string representing the castling rights of both sides in the game.
          */
-        static int fen_to_castling_rights(const std::string& fen);
+        static uint8_t fen_to_castling_rights(const std::string& fen);
 };
